+++ GUI(CLR)/GUI(CLR)/GUI(CLR).vcxproj
@@ -138,9 +138,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\src\classes.cpp" />
     <ClCompile Include="..\..\src\constant.cpp" />
     <ClCompile Include="..\..\src\core.cpp" />
+    <ClCompile Include="..\..\src\data_list.cpp" />
     <ClCompile Include="..\..\src\error.cpp" />
     <ClCompile Include="..\..\src\expression.cpp" />
     <ClCompile Include="..\..\src\filters.cpp" />
@@ -149,15 +149,18 @@
     <ClCompile Include="..\..\src\IO.cpp" />
     <ClCompile Include="..\..\src\math_builder.cpp" />
     <ClCompile Include="..\..\src\math_dummy.cpp" />
+    <ClCompile Include="..\..\src\math_vector.cpp" />
     <ClCompile Include="..\..\src\number.cpp" />
     <ClCompile Include="..\..\src\operations.cpp" />
     <ClCompile Include="..\..\src\variable.cpp" />
     <ClCompile Include="MainForm.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\src\classes.h" />
+    <ClInclude Include="..\..\src\common.h" />
+    <ClInclude Include="..\..\src\math_obj.h" />
     <ClInclude Include="..\..\src\constant.h" />
     <ClInclude Include="..\..\src\core.h" />
+    <ClInclude Include="..\..\src\data_list.h" />
     <ClInclude Include="..\..\src\error.h" />
     <ClInclude Include="..\..\src\expression.h" />
     <ClInclude Include="..\..\src\filters.h" />
@@ -166,6 +169,7 @@
     <ClInclude Include="..\..\src\IO.h" />
     <ClInclude Include="..\..\src\math_builder.h" />
     <ClInclude Include="..\..\src\math_dummy.h" />
+    <ClInclude Include="..\..\src\math_vector.h" />
     <ClInclude Include="..\..\src\number.h" />
     <ClInclude Include="..\..\src\operations.h" />
     <ClInclude Include="..\..\src\variable.h" />
