+++ src/math_vector.cpp
@@ -1,0 +1,16 @@
+#include "math_vector.h"
+namespace Project {
+	namespace Core {
+		using namespace std;
+
+
+		math_vector::math_vector()
+		{
+		}
+
+
+		math_vector::~math_vector()
+		{
+		}
+	}
+}
